Changed heading levels, quote style in drop-off #5, and typo in drop-off #7
Made Intro, Organ Recovery Process and CMS recommendations into H2, so it will show up in table of content

Made "Causes" and "Protential Fixes" to be bold and all caps

Made "Notes" to be H3, so it does NOT show up in table of contents. Made it bold and all caps.

Drop-off #3, changed to "MEDIUM/HIGH"

Under “Organ Recovery Process” - “Phase 3: Transport & Transplant” changed to be H3