Section 9 added for review (#51)
* Section 9 added for review

Hello!
This is the new section for the report. Will go in the last position in the left nav. Images and PDFs are already uploaded in the assets folder. It's not ready to go live yet, but we want it viewable so we can make final adjustments and send it off for review.

* Update 9_oversight-gaps-and-conflicts.md

* Update 9_oversight-gaps-and-conflicts.md

set nav to false

* Add files via upload

Updated file to include:
- thumbnails for appendix
- new links for appendix
- fixes to footnotes

* Update 9_oversight-gaps-and-conflicts.md

swapped in high res images for appendix

* Update 9_oversight-gaps-and-conflicts.md

Fixed instances of "58 OPOs" where it wasn't quoted or cited.

Co-authored-by: Matthew Crist <53876248+bloom-hondo@users.noreply.github.com>